Add files via upload
Deterministic Push Down Automaton!
Here is what is going on in the code (Python):

The structure of dpda.txt will be as follows:
Line 1: the states of the DPDA (separated by commas, if there is more than one state)
Line 2: the alphabet of the DPDA (separated by commas, if there is more than one symbol)
Line 3: the alphabet of the stack
Line 4: the starting state of the DPDA
Line 5: the final/accept states of the DPDA (separated by commas, if there is more than one accept state)
Line 6 and onward: the transition rules, where each rule takes the form a,b,c,d,e (where being in state ‘a’ and reading symbol ‘b’ while popping ‘c’ from the top of the stack transitions to new state ‘d’ and pushes ‘e’ on the top of the stack).

In addition to the given alphabet, all DPDAs may contain empty-string transitions for either the input character, the symbol at the top of the stack, or both (we will use ‘@’ to represent this). You can use any symbol for the bottom of the stack, but the provided examples I will give you use a ‘$’. After your program constructs a DPDA, read in another file called input.txt, where each line in the file is a string that will run on your DPDA. Both dpda.txt and input.txt should be assumed to be in the same directory as your Python program. Simulate each string in input.txt using your DPDA, and write the output (accept or reject) into a new file called output.txt (also in the same directory as your Python program, and each result on a new line.